gas_genie/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
fireworks-ai>=0.15.12
langchain-core>=0.1.0
sentient-agent-framework>=0.1.0 